requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
mysql-connector-python==8.0.27
protobuf==3.19.1
python-dotenv==0.19.2
requests==2.26.0
urllib3==1.26.7
